+++ public/assets/i18n/common/hr.json
@@ -25,7 +25,7 @@
   "sudjerica": "Perilica posuđa",
   "vesmasina": "Perilica rublja",
   "app1Text": "Apartman 1 se nalazi u prizemlju. Kategoriziran je kao 4+1, što znači da se u njega može smjestiti ukupno 5 osoba. Sastoji se od dnevne sobe sa krevetom na razvlačenje i blagovaonicom, dvije spavaće sobe (jedna sa bračnim krevetom, a druga sa dva odvojena kreveta), kuhinje, kupaonice i natkrivene terase. Od dodatnih sadržaja apartman ima klimu, Wi-Fi, perilicu posuđa, perilicu rublja, mikrovalnu pećnicu, pećnicu, kuhalo za vodu te TV uređaj. Svi dodatni sadržaji su besplatni.",
-  "app2Text": "Apartman 2 se nalazi u prizemlju, no njemu se prilazi sa zapadne strane kuće. Kategoriziran je kao 2+2, što znači da se u njega može smjestiti ukupno 4 osobe. Zbog manje površine, ali i mirnog okruženja idealan je za parove. Sastoji se od dnevne sobe sa krevetom na razvlačenje, jedne spavaće sobe sa dva odvojena kreveta, kuhinjom, kupaonicom te terasom. Od dodatnih sadržaja apartman ima klimu, Wi-Fi, mikrovalnu pećnicu, kuhalo za vodu i TV uređaj. Svi dodatni sadržaji su besplatni",
+  "app2Text": "Apartman 2 se nalazi u prizemlju, no njemu se prilazi sa zapadne strane kuće. Kategoriziran je kao 2+2, što znači da se u njega može smjestiti ukupno 4 osobe. Zbog manje površine, ali i mirnog okruženja idealan je za parove. Sastoji se od dnevne sobe sa krevetom na razvlačenje, jedne spavaće sobe sa dva odvojena kreveta, kuhinjom, kupaonicom te terasom. Od dodatnih sadržaja apartman ima klimu, Wi-Fi, mikrovalnu pećnicu, kuhalo za vodu i TV uređaj. Svi dodatni sadržaji su besplatni.",
   "app3Text": "Apartman 3 je smješten na katu kuće. Kategoriziran je kao 2+2, što znači da se u njega može smjestiti ukupno 4 osobe. Sastoji se od dnevne sobe koja je povezana sa blagovaonicom i kuhinjom, jedne velike spavaće sobe (koja sadrži jedan bračni krevet i dva odvojena kreveta), kupaonice i otvorene terase. Od dodatnih sadržaja apartman ima klimu, Wi-Fi, perilicu posuđa, perilicu rublja, mikrovalnu pećnicu, pećnicu, kuhalo za vodu te TV uređaj. Svi dodatni sadržaji su besplatni.",
   "kontakt": "Kontakt",
   "vlasnik": "Vlasnik: ",
