+++ src/Components/Apartments.css
@@ -1,6 +1,6 @@
 .basicInfo {
   width: 88%;
-  margin: 10% auto 0 auto;
+  margin: 5% auto 5% auto;
   display: flex;
   border: 1px solid rgba(255, 255, 255, 1);
   background-color: #ffffff;
@@ -32,7 +32,7 @@
   padding: 0.5rem;
   box-shadow: 0px 8px 24px 0px rgba(0, 0, 0, 0.25);
   border: none;
-  margin: 1rem auto 0 auto;
+  margin: 1rem auto 1rem auto;
   height: 5.25vh;
   font-family: Poppins;
   font-size: 1.15rem;
@@ -62,4 +62,34 @@
 .Description {
   width: 88%;
   margin: 0.4rem auto;
+}
+@media (min-width: 850px) and (max-width: 1070px) {
+  .basicInfo {
+    flex-direction: column;
+    height: 3.5rem;
+  }
+  .Half {
+    border-right: none;
+    border-bottom: 1px solid #276798;
+    width: 90%;
+    margin: 0.2rem;
+  }
+  .Half:last-child {
+    border-bottom: none;
+  }
+}
+@media (max-width: 330px) {
+  .basicInfo {
+    flex-direction: column;
+    height: 3.5rem;
+  }
+  .Half {
+    border-right: none;
+    border-bottom: 1px solid #276798;
+    width: 90%;
+    margin: 0.2rem;
+  }
+  .Half:last-child {
+    border-bottom: none;
+  }
 }
